requirements.txt
feedparser==6.0.10
requests==2.31.0
selenium==4.15.0
openai-whisper==20231117
pydub==0.25.1
beautifulsoup4==4.12.2
python-dotenv==1.0.0
